Add more ways to provide the CDFF to the DFN template generator
The DFN template generator needs the path to the CDFF (Core+Support). It looks
for it on the command line and defaults to the fixed string "CDFF" if it isn't
found.

This commit uses the cdff_dev.path module to make the DFN template generator
look for the path to the CDFF on the command line (--cdffpath), in the
environment variable CDFFPATH, and in a hint file in the current directory,
in this order of precedence.