Add more ways to provide the CDFF to the DFPC template generator
Exactly the same as 35a19f2b2348c4d72819da6e7f0870f8e54ff20d, but for the DFPC
template generator rather than the DFN template generator.

The DFPC template generator needs the path to the CDFF (Core+Support). It looks
for it on the command line and defaults to the fixed string "CDFF" if it isn't
found.

This commit uses the cdff_dev.path module to make the DFPC template generator
look for the path to the CDFF on the command line (--cdffpath), in the
environment variable CDFFPATH, and in a hint file in the current directory,
in this order of precedence.