Use an explicit Jinja2 environment to render Jinja2 template files
Using the jinja2.Template() constructor to directly create a Jinja2 template
object from a file in Jinja2 markup language also automatically creates an
anonymous Jinja2 environment object.

Unfortunately, it is impossible to customize that environment object (e.g.
control whitespace trim), because the object is already initialized, and it is
also unsafe to add custom filters to it (e.g. capitalize first character of a
string), because a template is already loaded.

To make that possible, this commit explicitly creates a named Jinja2
environment object, then uses that environment object to create the Jinja2
template object from the file in Jinja2 markup language.

Source: main documentation: http://jinja.pocoo.org/docs/2.10/api

Also, argument changed from template to template_name to emphasize that it is
not the Jinja2 template object.