Make cdffpath a mandatory argument of the DFN and DFPC code generators
Certain functions in the DFN and DFPC code generators default to "CDFF" as the
path to the CDFF, however there is little ground to assume that there will be
a subdirectory (or a symlink) named "CDFF" in whatever directory happens to be
the current directory.

This commit refactors the argument lists of the functions check_cdffpath,
write_dfn, and write_dfpc to make cdffpath a mandatory argument. Calls to these
functions in the DFN and DFPC template generators and in the unit tests are
adjusted accordingly.

The path to the CDFF can be given on the command line (--cdffpath), in the
environment variable CDFFPATH, and in the hint file cdffpath in the current
directory, in this order of precedence, see !52.