Add the University of Strathclyde as a contributor
This is the "Corporate logo (with tab)". Brand rules require using the
"Full-colour logo (without tab)" for "partnership project materials that
feature a range of logos and are not led by the University of Strathclyde".
This may require paperwork and be subject to authorization. The "Corporate
logo (with tab)" will act as a placeholder in the meantime.

See also #15 and https://www.strath.ac.uk/brand/